Fix client connecting via gossip seeds
- This fixes the issue reported on the list in the "connection to cluster
  using gossip seeds fails (client api)" thread.
- ClusterSettingsBuilder now forces you to choose first between DNS or
  Gossip Seed configuration and then presents only the correct
  configuration points.
- This is a breaking API change!
