+++ src/EventStore/EventStore.ClientAPI/EventStore.ClientAPI.csproj
@@ -84,6 +84,8 @@
   <ItemGroup>
     <Compile Include="AllCheckpoint.cs" />
     <Compile Include="AllEventsSlice.cs" />
+    <Compile Include="Core\GossipSeedEndPointDiscoverer.cs" />
+    <Compile Include="Core\NodeEndPoints.cs" />
     <Compile Include="StreamCheckpoint.cs" />
     <Compile Include="ClientAuthenticationFailedEventArgs.cs" />
     <Compile Include="ClientErrorEventArgs.cs" />
